+++ actix-multipart/src/extractor.rs
@@ -33,6 +33,7 @@ use crate::server::Multipart;
 impl FromRequest for Multipart {
     type Error = Error;
     type Future = Result<Multipart, Error>;
+    type Config = ();
 
     #[inline]
     fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {