+++ actix-session/src/lib.rs
@@ -175,6 +175,7 @@ impl Session {
 impl FromRequest for Session {
     type Error = Error;
     type Future = Result<Session, Error>;
+    type Config = ();
 
     #[inline]
     fn from_request(req: &HttpRequest, _: &mut Payload) -> Self::Future {