+++ src/request.rs
@@ -266,6 +266,7 @@ impl Drop for HttpRequest {
 /// }
 /// ```
 impl FromRequest for HttpRequest {
+    type Config = ();
     type Error = Error;
     type Future = Result<Self, Error>;