+++ awc/src/ws.rs
@@ -306,9 +306,7 @@ impl WebsocketsRequest {
                         }
                     } else {
                         log::trace!("Invalid connection header: {:?}", conn);
-                        return Err(WsClientError::InvalidConnectionHeader(
-                            conn.clone(),
-                        ));
+                        return Err(WsClientError::InvalidConnectionHeader(conn.clone()));
                     }
                 } else {
                     log::trace!("Missing connection header");